+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* **Tweak** - Resaweb integration, add suffix parameter
+
 ### 1.1.19: Janvuary 14th, 2020
 * **Fix** - Queried objets not correctly obtained in Loop Elements in Elementor Theme Builder (new fix)